ChatGPT/back_ChatGPT/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pyodbc
python-dotenv